Docs: Clarify adaptive playback requirement for video codecs
Android manufactures asked whether the current adaptive playback
requirement between video codecs apply to ALL available codecs
reported by the device through MediaCodec.

This update is to clarify that this is indeed the case, so that
we can assure adaptive playback works properly on all video apps
regardless of which video stream they are playing.

Bug: 20959915

Change-Id: I081df97ac8e2495c1263523998bf76eea6cccb88