Make use of keystore and keymaster names consistent. Also fix bugs.
The documentation often says "keymaster" where it shoud say "keystore", and
when it correctly says "keymaster" it often fails to distinguish between
the keymaster HAL and the underlying secure hardware.  In addition,
there are a number of content errors identified by partners in the
process of implementing the keymaster specification. This CL updates the
documentation to correct both of those issues.

Change-Id: Iba4d75493c5a419a1ddf7aeeafc13a7dd53f6b46