CDD: Pre-installed system apps may write to secondary ext. storage Section: 7.6.2. Application Shared Storage
Clarify that pre-installed system previliged apps, for example the
default out-of-box photo gallery app, may have write access to the
secondary external storage.

The original language may be misunderstood that, even pre-installed
Android applications that functions as a core function of the platform
MUST NOT be allowed from having write-access to the secondary external
storage.

Bug: 27477907
Change-Id: Ib9680011f72052fd4225206894e4475d3113dd01