CDD: Remove redundant requirement on UMS/MTP from 7.7 Section: 7.7. USB
Section 7.6.2 is already describing the requirement of when and
how the access to the application shared storage has to be provided
for the USB host machine.

Over time this requirement in 7.6.2 has been made more flexible in
terms of which technical protocol is being used for the practice of
moving around media files from and into the application shared
storage of the device implementation. This decision reflects the
increased usage of other alternative methods using local or cloud
based over-the-air sync of the media files.

Bug: 25671486
Change-Id: I2be2b1344dc207c678f1e4d0a0abc51a49979414