CDD: Fixing typos in the USB section Section: 7.7. USB
Corrected the USB Battery Charging Specification URL, as it was
previously referring to the testing specification.

Corrected the recommended on-device USB port type to properly refer
to type-B instead of type-A for when the device implementation has
USB peripheral mode support.

Removed the recommendation to center the USB peripheral port.

Bug: 24936011
Bug: 24935017

Change-Id: I5e2a8b62dcca3104443335c23eac63bfad6fa7ec