CDD: Clarify dynamic switching of video resolution is up to max value per codec Section: 5.3. Video Decoding
The "maximum resolution supported on the device" from the following language is
referring to the per codec maximum resolution, not necessarily the screen
resolution of the device itself as being reported by Display.getSupportedModes().

"... MUST support dynamic video resolution and frame rate switching through the
standard Android APIs ... up to the maximum resolution supported on the device"

This was not clear in the original language, but clarifed further by explicitly
saying "maximum resolution supported by each codec".

Also, improved the "is not smaller than" to "equal or greater than" for better
readibility.

Bug: 27301996
Change-Id: I3f9bbd51e7b3abdfd41fa97e5b54e135fd036bf3