CDD: Call out that AOA protocol based connection MUST work w/o changing the USB mode Section: 7.7. USB
The requirement to implement the AOA specification, implicitly
includes the requirement to support the capability to establish
an AOA protocol based communication between the Android device
implementation and an AOA protocol based accessory as described
in the following page on source.android.com.

http://source.android.com/devices/accessories/aoa.html

This CDD change is to call this out more explicitly to avoid
further confusion.

Bug: 25487364
Change-Id: I105cf1c9ea2e2e6b679f34ff194085da651a942d